refactor: Extract the scroll provifer to a dedicated directory
