+++ hooks/index.tsx
@@ -2,7 +2,6 @@ export { useDebouncedCallback } from './use-debounced-callback'
 export { useHeadings } from './use-headings'
 export { useNewsletter } from './use-newsletter'
 export { useReducedMotion } from './use-reduced-motion'
-export { ScrollProvider, useScroll } from './use-scroll'
 export { useScrollToHeading } from './use-scroll-to-heading'
 export { useTheme } from './use-theme'
 export { useDocumentDimensions } from './use-document-dimensions'
