+++ providers/index.tsx
@@ -1,0 +1,1 @@
+export { ScrollProvider, useScroll } from './scroll-provider'
